fix(critical): sortOrder Long, scrittura atomica con backup temp, errore esplicito su caricamento
- sortOrder: da Int a Long (evita troncamento timestamp)
- save(): scrivi temp prima, poi criptato, elimina temp
- loadWithResult(): sealed class LoadResult con Success/Empty/Migrated/Error
- Migrazione temp file da crash precedente

+++ app/src/main/java/com/card/fidelybar/FidelyBarViewModel.kt
@@ -65,7 +65,7 @@ class FidelyBarViewModel(app: Application) : AndroidViewModel(app) {
             logoUrl = logoUrl,
             logoBorderWhite = logoBorderWhite,
             logoBorderSize = logoBorderSize,
-            sortOrder = now.toInt(),
+            sortOrder = now,
             createdAt = now,
             updatedAt = now
         )
